conver to long

+++ src/OSPSuite.SimModelSolver_CVODES/include/SimModelSolver_CVODES/SimModelSolver_CVODES.h
@@ -54,7 +54,7 @@ private:
 	realtype _relTol_CVODE;
 
 	// internal step counter
-	int _step;
+	long _step;
 
 	//fill CVOde specific Solver options in datatypes required by CVODE
 	void FillSolverOptions(void);
